+++ src/api/web/tracks.rs
@@ -21,7 +21,7 @@ use aoide_storage::{
         track::{TrackAlbums, TrackTags, Tracks, TracksResult},
         CountTagFacetsParams, CountTagsParams, CountTrackAlbumsParams, LocateTracksParams,
         Pagination, PaginationLimit, PaginationOffset, ReplaceTracksParams, ReplacedTracks,
-        SearchTracksParams, StringPredicate, TagCount, TagFacetCount, UriPredicate,UriRelocation,
+        SearchTracksParams, StringPredicate, TagCount, TagFacetCount, UriPredicate, UriRelocation,
     },
     storage::track::TrackRepository,
 };
@@ -205,13 +205,9 @@ impl TracksHandler {
         query_params: TracksQueryParams,
         uri_relocations: impl IntoIterator<Item = UriRelocation>,
     ) -> impl Future<Item = impl warp::Reply, Error = warp::reject::Rejection> {
-        relocate_tracks(
-            &self.db,
-            query_params.collection_uid,
-            uri_relocations,
-        )
-        .map(|()| StatusCode::NO_CONTENT)
-        .map_err(warp::reject::custom)
+        relocate_tracks(&self.db, query_params.collection_uid, uri_relocations)
+            .map(|()| StatusCode::NO_CONTENT)
+            .map_err(warp::reject::custom)
     }
 
     pub fn handle_albums_count(
@@ -389,8 +385,13 @@ fn relocate_tracks(
     let repository = TrackRepository::new(&*pooled_connection);
     future::result(pooled_connection.transaction::<_, Error, _>(|| {
         for uri_relocation in uri_relocations {
-            let locate_params = LocateTracksParams {
-                uri: StringPredicate::StartsWith(uri_relocation.old_prefix.clone()),
+            let locate_params = match &uri_relocation.predicate {
+                UriPredicate::Prefix(uri_prefix) => LocateTracksParams {
+                    uri: StringPredicate::StartsWith(uri_prefix.to_owned()),
+                },
+                UriPredicate::Exact(uri) => LocateTracksParams {
+                    uri: StringPredicate::Equals(uri.to_owned()),
+                },
             };
             let tracks = repository.locate_entities(
                 collection_uid.as_ref(),
@@ -398,16 +399,21 @@ fn relocate_tracks(
                 locate_params,
             )?;
             log::debug!(
-                "Found {} track(s) that match {} as candidates for relocating",
+                "Found {} track(s) that match {:?} as candidates for relocating",
                 tracks.len(),
-                uri_relocation.old_prefix,
+                uri_relocation.predicate,
             );
             for track in tracks {
                 let format = track.format;
                 let mut track: TrackEntity = track.deserialize()?;
-                let relocation_count = track
-                    .body_mut()
-                    .relocate_source_by_uri_prefix(&uri_relocation.old_prefix, &uri_relocation.new_prefix);
+                let relocation_count = match &uri_relocation.predicate {
+                    UriPredicate::Prefix(uri_prefix) => track
+                        .body_mut()
+                        .relocate_source_by_uri_prefix(uri_prefix, &uri_relocation.replacement),
+                    UriPredicate::Exact(uri) => track
+                        .body_mut()
+                        .relocate_source_by_uri(uri, &uri_relocation.replacement),
+                };
                 if relocation_count > 0 {
                     log::debug!(
                         "Updating track {} after relocating {} source(s)",